move multi into engine [pr]
